pyproject.toml
[project]
name = "mkts-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "millify>=0.1.1",
    "pandas>=2.3.0",
    "requests>=2.32.4",
    "requests-oauthlib>=2.0.0",
    "sqlalchemy-libsql>=0.2.0",
    "sqlalchemy-orm>=1.2.10",
    "colorlog>=6.8.0",
    "matplotlib>=3.10.5",
    "seaborn>=0.13.2",
    "plotly>=6.2.0",
    "httpx[http2]>=0.28.1",
    "aiolimiter>=1.2.1",
    "backoff>=2.2.1",
    "sqlalchemy>=2.0.43",
    "libsql>=0.1.11",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mkts_backend"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
mkts-backend = "mkts_backend.cli:main"
